7B_custom_airflow_operators/cde_setup/requirements.txt
apache-airflow-providers-http
pandas
llama-index
custom-airflow-operator-superellipse==0.0.6